+++ UniversalDeviceToolkit.Lib/Extensions/TaskExtensions.cs
@@ -9,9 +9,19 @@ public static class TaskExtensions
 {
     public static ValueTask AsValueTask(this Task task) => new(task);
 
-    public static Task<T?> OrNullIfException<T>(this Task<T> task) where T : struct
+    public static async Task<T?> OrNullIfException<T>(this Task<T> task) where T : struct
     {
-        return task.ContinueWith(t => t.IsCompletedSuccessfully ? (T?)t.Result : null);
+        try
+        {
+            return await task.ConfigureAwait(false);
+        }
+        catch (Exception ex)
+        {
+            if (Log.Instance.IsTraceEnabled)
+                Log.Instance.Trace("Task completed with an exception; returning null.", ex);
+
+            return null;
+        }
     }
 
     public static void Forget(this Task task, string operationName)